+++ lib/src/generator/serializer_generator.dart
@@ -1,5 +1,6 @@
 import 'package:entity_serializer/entity_serializer.dart';
 import 'package:entity_serializer/src/model/serializer.dart';
+import 'package:entity_serializer/src/model/specialization.dart';
 import 'package:entity_serializer/src/model/specialization_convert.dart';
 import 'package:xml/xml.dart';
 
@@ -25,8 +26,7 @@ class SerializerGenerator {
           _parseKeepNode(child, result);
           break;
         default:
-          throw Exception(
-              "Unknown node '${node.name.toString()}' in serializer '$name'");
+          throw Exception("Unknown node '${node.name.toString()}' in serializer '$name'");
       }
     }
 
@@ -36,10 +36,8 @@ class SerializerGenerator {
   static void _parseSpecializationNode(XmlElement node, Serializer serializer) {
     final serFunc = attrValue(node, "serialization");
     final deserFunc = attrValue(node, "deserialization");
-    if ((serFunc == null || serFunc.isEmpty) &&
-        (deserFunc == null || deserFunc.isEmpty)) {
-      throw Exception(
-          "Both 'serialization' and 'deserialization' attributes are empty"
+    if ((serFunc == null || serFunc.isEmpty) && (deserFunc == null || deserFunc.isEmpty)) {
+      throw Exception("Both 'serialization' and 'deserialization' attributes are empty"
           " in node '${node.name}', this make no sense.");
     }
     final importPath = reqAttrValue(node, "import");
@@ -47,17 +45,28 @@ class SerializerGenerator {
       throw Exception("Attribute 'import' is empty in node '${node.name}'");
     }
     final type = reqAttrValue(node, "type");
-    serializer.addSpecialization(
-        type,
-        SpecializationConvert(
-            inType: type,
-            importPath: importPath,
-            serializationFunc: serFunc,
-            deserializationFunc: deserFunc,
-            outType: attrValue(node, "outType")));
+    final specialization = SpecializationConvert(
+        inType: type,
+        importPath: importPath,
+        serializationFunc: serFunc,
+        deserializationFunc: deserFunc,
+        outType: attrValue(node, "outType"));
+    _collectFieldsMapping(specialization, node);
+    serializer.addSpecialization(type, specialization);
   }
 
-  static void _parseKeepNode(XmlElement node, Serializer serializer) =>
-      serializer.addSpecialization(
-          reqAttrValue(node, "type"), SpecializationKeep());
+  static void _parseKeepNode(XmlElement node, Serializer serializer) {
+    final specialization = SpecializationKeep();
+    _collectFieldsMapping(specialization, node);
+    serializer.addSpecialization(reqAttrValue(node, "type"), specialization);
+  }
+
+  static void _collectFieldsMapping(Specialization specialization, XmlElement node) {
+    node.attributes.where((a) => a.name.toString().startsWith("_")).forEach((a) {
+      final key = a.name.toString().substring(1);
+      if (key.isNotEmpty) {
+        specialization.aliases.putIfAbsent(key, () => a.value);
+      }
+    });
+  }
 }