+++ lib/src/model/serializer.dart
@@ -9,11 +9,9 @@ class Serializer {
 
   Serializer({required this.name});
 
-  void addSpecialization(String type, Specialization spec) =>
-      _specializations[type] = spec;
+  void addSpecialization(String type, Specialization spec) => _specializations[type] = spec;
 
-  bool hasSpecialization(Field field) =>
-      _specializations.containsKey(_getFieldType(field));
+  bool hasSpecialization(Field field) => _specializations.containsKey(_getFieldType(field));
 
   String handleSerialization(Field field, String variable) =>
       _specializations[_getFieldType(field)]!.processSerialization(variable);
@@ -21,8 +19,7 @@ class Serializer {
   String handleDeserialization(Field field, String variable) =>
       _specializations[_getFieldType(field)]!.processDeserialization(variable);
 
-  String _getFieldType(Field field) =>
-      "${field.type}${field.isOptional ? "?" : ""}";
+  String _getFieldType(Field field) => "${field.type}${field.isOptional ? "?" : ""}";
 
   void collectImports(ImportWriter collector) {
     for (final spec in _specializations.values) {
@@ -47,5 +44,12 @@ class Serializer {
     for (final ent in template._specializations.entries) {
       _specializations.putIfAbsent(ent.key, () => ent.value);
     }
+  }
+
+  String nameFor(Field field) {
+    final spec = _specializations[_getFieldType(field)];
+    String? result = spec?.aliases[field.name];
+    result ??= field.name;
+    return result;
   }
 }