+++ lib/src/model/specialization.dart
@@ -1,4 +1,6 @@
 abstract class Specialization {
+  Map<String, String> aliases = {};
+
   String? get getImportPath;
   String processSerialization(String data);
   String processDeserialization(String data);